+++ app/__init__.py
@@ -4,7 +4,7 @@
 from flask import Flask
 
 # Initialize the app
-app = Flask(__name__, instance_relative_config=True)
+app = Flask(__name__)
 
 # Load the config file
 app.config.from_object('config')
