+++ app/apis/__init__.py
@@ -1,0 +1,2 @@
+# app/apis/__init__.py
+"""Weconnect api entry point"""
